+++ ReleaseNotes.md
@@ -14,6 +14,7 @@ Improved resource handling for WavPack and FFMpeg requrires reimport of the libr
 - Filter resources by mimetype instead of extension
 - Artwork support for dsf files with taglib 2.0
 - "PC Directory" can be configured with `<boxlayout>`
+- MySql: Configuration of engine, charset and collation before first run
 
 ### FIXES